% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{constructGraph}
\alias{constructGraph}
\title{constructGraph}
\usage{
constructGraph(data)
}
\arguments{
\item{data}{Tibble; Expected columns are "from" and "to", indicating the
edges and "dist" indicating the distance between the samples.}
}
\description{
Constructs a graph from a data frame or tibble.
}
\author{
Jason T. Serviss
}
\keyword{internal}
