% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{distInPCAspace}
\alias{distInPCAspace}
\title{distInPCAspace}
\usage{
distInPCAspace(pca, pcs)
}
\arguments{
\item{pca}{Matrix; Principal components as columns, samples as rows.}

\item{pcs}{Integer; Length 1 vector indicating the max principal component to
retain.}
}
\description{
Calculates Euclidean distance in PCA space.
}
\author{
Jason T. Serviss
}
\keyword{internal}
