% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{louvain}
\alias{louvain}
\title{louvain}
\usage{
louvain(g, weights.col)
}
\arguments{
\item{g}{tbl_graph; A tidygraph graph.}

\item{weights.col}{Bare word; indicating the column name of the column
including the weights.}
}
\description{
Runs the louvain community detection algorithm on the input graph.
}
\author{
Jason T. Serviss
}
\keyword{internal}
